Рефакторинг

+++ app/src/main/java/com/vgb3/kstore/data/model/input/VaultSummaryData.kt
@@ -8,6 +8,6 @@ data class VaultSummaryData(
     val createdAt: Long,
     val updatedAt: Long,
     val isFavorite: Boolean,
-    val categoryId: Int?,
+    val categoryId: Long?,
     val type: VaultItemType,
 )
